write_csv/requirements.txt
pika==1.1.0
PyMySQL==0.9.3
mysql-connector-python==8.0.19